requirements.txt
Django>=4.2
djangorestframework
djangorestframework-simplejwt
drf-yasg
psycopg2-binary
python-dotenv
